+++ CHANGELOG.md
@@ -37,8 +37,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   support restricting the search space via `active_values`, while `values` continue to 
   identify allowed measurement inputs
 - `Campaign.get_acquisition_function` and `BayesianRecommender.get_acquisition_function`
-  convenience methods
-- `Campaign.acquisition_values` convenience method for computing acquisition values
+  convenience methods for retrieving the underlying acquisition function
+- `AcquisitionFunction.evaluate` convenience method for computing acquisition values
+- `acquisition_values` and `joint_acquisition_value` convenience methods to
+  `Campaign` and `BayesianRecommender` for computing acquisition values
 
 ### Changed
 - Acquisition function indicator `is_mc` has been removed in favor of new indicators